Extract getFormattedDateTime to an util file
