+++ packages/nextjs/app/admin/_components/SubmissionComments.tsx
@@ -5,18 +5,9 @@ import { useRouter } from "next/navigation";
 import { useMutation } from "@tanstack/react-query";
 import { Address } from "~~/components/scaffold-eth";
 import { Submission } from "~~/services/database/repositories/submissions";
+import { getFormattedDateTime } from "~~/utils/date";
 import { postMutationFetcher } from "~~/utils/react-query";
 import { notification } from "~~/utils/scaffold-eth";
-
-export function getFormattedDateTime(date: Date) {
-  const month = String(date.getMonth() + 1).padStart(2, "0");
-  const day = String(date.getDate()).padStart(2, "0");
-  const year = date.getFullYear();
-  const hours = date.getHours();
-  const minutes = String(date.getMinutes()).padStart(2, "0");
-
-  return `${month}/${day}/${year} ${hours}:${minutes}`;
-}
 
 export const SubmissionComments = ({ submission }: { submission: Submission }) => {
   const [newComment, setNewComment] = useState("");
