+++ packages/nextjs/utils/date.ts
@@ -1,0 +1,9 @@
+export function getFormattedDateTime(date: Date) {
+  const month = String(date.getMonth() + 1).padStart(2, "0");
+  const day = String(date.getDate()).padStart(2, "0");
+  const year = date.getFullYear();
+  const hours = date.getHours();
+  const minutes = String(date.getMinutes()).padStart(2, "0");
+
+  return `${month}/${day}/${year} ${hours}:${minutes}`;
+}
